+++ ExilenceClient/src/app/authorize/components/income/income.component.ts
@@ -9,6 +9,8 @@ import { Party } from '../../../shared/interfaces/party.interface';
 import { Player } from '../../../shared/interfaces/player.interface';
 import { PartyService } from '../../../shared/providers/party.service';
 import { SettingsService } from '../../../shared/providers/settings.service';
+import { ElectronService } from '../../../shared/providers/electron.service';
+import { Router } from '@angular/router';
 
 
 @Component({
@@ -28,8 +30,9 @@ export class IncomeComponent implements OnInit, OnDestroy {
 
   public isHidden = false;
   public visible = true;
-
   public isSummary = false;
+  public foundPlayer: Player;
+
   private selectedPlayerSub: Subscription;
   private partySubscription: Subscription;
   private selectedFilterValueSub: Subscription;
@@ -50,7 +53,16 @@ export class IncomeComponent implements OnInit, OnDestroy {
   constructor(
     private partyService: PartyService,
     private settingService: SettingsService,
+    private router: Router
   ) {
+  }
+
+  anyPlayerSnapshots() {
+    return this.party.players.find(x =>
+      x.netWorthSnapshots !== undefined
+      && x.netWorthSnapshots.length > 0
+      && x.netWorthSnapshots[0].timestamp > 0
+    ) !== undefined;
   }
 
   ngOnInit() {
@@ -68,9 +80,9 @@ export class IncomeComponent implements OnInit, OnDestroy {
       // update the graph every minute, to update labels
       this.interval = setInterval(() => {
         this.dateData = [];
-        const foundPlayer = this.party.players.find(x => x.character.name === this.partyService.selectedFilterValue);
-        if (this.partyService.selectedFilterValue !== '0' && foundPlayer !== undefined) {
-          this.updateGraph(foundPlayer);
+        this.foundPlayer = this.party.players.find(x => x.character.name === this.partyService.selectedFilterValue);
+        if (this.partyService.selectedFilterValue !== '0' && this.foundPlayer !== undefined) {
+          this.updateGraph(this.foundPlayer);
         } else {
           this.party.players.forEach(p => {
             if (p.netWorthSnapshots !== null) {
@@ -88,9 +100,9 @@ export class IncomeComponent implements OnInit, OnDestroy {
           this.party = party;
 
           // update values for entire party, or a specific player, depending on selection
-          const foundPlayer = this.party.players.find(x => x.character.name === this.partyService.selectedFilterValue);
-          if (this.partyService.selectedFilterValue !== '0' && foundPlayer !== undefined) {
-            this.updateGraph(foundPlayer);
+          this.foundPlayer = this.party.players.find(x => x.character.name === this.partyService.selectedFilterValue);
+          if (this.partyService.selectedFilterValue !== '0' && this.foundPlayer !== undefined) {
+            this.updateGraph(this.foundPlayer);
           } else {
             this.party.players.forEach(p => {
               if (p.netWorthSnapshots !== null) {
@@ -107,9 +119,9 @@ export class IncomeComponent implements OnInit, OnDestroy {
           this.dateData = [];
 
           // update values for entire party, or a specific player, depending on selection
-          const foundPlayer = this.party.players.find(x => x.character.name === this.partyService.selectedFilterValue);
-          if (this.partyService.selectedFilterValue !== '0' && foundPlayer !== undefined) {
-            this.updateGraph(foundPlayer);
+          this.foundPlayer = this.party.players.find(x => x.character.name === this.partyService.selectedFilterValue);
+          if (this.partyService.selectedFilterValue !== '0' && this.foundPlayer !== undefined) {
+            this.updateGraph(this.foundPlayer);
           } else {
             this.party.players.forEach(p => {
               if (p.netWorthSnapshots !== null) {
@@ -135,6 +147,10 @@ export class IncomeComponent implements OnInit, OnDestroy {
     if (this.interval) {
       clearInterval(this.interval);
     }
+  }
+
+  goToSettings() {
+    this.router.navigate(['/authorized/settings']);
   }
 
   hideGraph() {