+++ ExilenceClient/src/app/authorize/components/server-message-dialog/server-message-dialog.component.ts
@@ -22,7 +22,7 @@ export class ServerMessageDialogComponent implements OnInit {
   onNoClick(): void {
     this.dialogRef.close();
   }
-  disableTooltips() {
+  close() {
     this.dialogRef.close();
   }
   ngOnInit() {