+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.9.14] - 2019-XX-XX
+## [0.9.14] - 2019-01-29
 ### Added
 - Added support for pricing corrupted items (uniques and gems)
 - Added new functionality for "leaders"